requirements.txt
-e git+https://github.com/django/django.git@259259a819a12a9ecfd0448c646899fad78e23e0#egg=Django-origin/stable/1.8.x
django-environ==0.3.0
requests==2.5.1
psycopg2==2.6
python-dateutil==2.4.0
pytz==2014.10
six==1.9.0
beautifulsoup4==4.3.2
